+++ ddtrace/_trace/utils_redis.py
@@ -45,20 +45,3 @@ def _set_span_tags(
         for attr in ("command_stack", "_command_stack"):
             if hasattr(instance, attr):
                 span.set_metric(redisx.PIPELINE_LEN, len(getattr(instance, attr)))
-
-
-@contextmanager
-def _instrument_redis_execute_async_cluster_pipeline(pin, config_integration, cmds, instance):
-    cmd_string = resource = "\n".join(cmds)
-    if config_integration.resource_only_command:
-        resource = "\n".join([cmd.split(" ")[0] for cmd in cmds])
-
-    with core.context_with_data(
-        "redis.async_cluster_pipeline.execute",
-        span_name=schematize_cache_operation(redisx.CMD, cache_provider=redisx.APP),
-        resource=resource,
-        service=trace_utils.ext_service(pin, config_integration),
-        span_type=SpanTypes.REDIS,
-    ) as span:
-        core.dispatch("redis.execute_pipeline", [ctx, pin, config_integration, None, instance, cmd_string])
-        yield span
