+++ ddtrace/contrib/internal/redis/asyncio_patch.py
@@ -1,6 +1,6 @@
 from ddtrace import config
-from ddtrace._trace.utils_redis import _instrument_redis_execute_async_cluster_pipeline
 from ddtrace.contrib.internal.redis_utils import _instrument_redis_cmd
+from ddtrace.contrib.internal.redis_utils import _instrument_redis_execute_async_cluster_pipeline
 from ddtrace.contrib.internal.redis_utils import _instrument_redis_execute_pipeline
 from ddtrace.contrib.internal.redis_utils import _run_redis_command_async
 from ddtrace.internal.utils.formats import stringify_cache_args
